projectalphav1/requirements.txt
asgiref==3.9.1
Django==5.2.5
psycopg2-binary>=2.9.10
sqlparse==0.5.3
Faker==37.5.3
factory_boy>=3.3.0,<4
tzdata==2025.2
python-dotenv==1.0.1
dj-database-url==2.1.0
Pillow>=11.1.0
python-dateutil>=2.8.0,<3

# API framework and CORS
djangorestframework>=3.15,<4
django-cors-headers>=4.4,<5
django-filter>=24.0,<25

# Field-level encryption for PII
# Docs: https://github.com/jazzband/django-fernet-encrypted-fields
# Compatible with Django 4.2+ (works with 5.2)
django-fernet-encrypted-fields>=0.3.1,<1

# Modern Django Admin Interface
# Docs: https://github.com/fabiocaccamo/django-admin-interface
django-admin-interface>=0.28,<1

# AI/ML APIs
# Anthropic Python SDK for AI summaries and ETL column mapping (deprecated, use Gemini instead)
anthropic>=0.30,<1

# Google Generative AI SDK for Gemini 2.5 Flash (recommended for document vision)
# Docs: https://ai.google.dev/gemini-api/docs/quickstart?lang=python
google-generativeai>=0.8.0,<1

# Geocodio official Python client (latest as of now is 0.2.x)
geocodio-library-python>=0.2.0,<1

# Data processing and ETL
# Docs: https://numpy.org/doc/
numpy>=1.24.0
# Financial calculations (IRR, NPV)
# Docs: https://numpy-financial.readthedocs.io/
numpy-financial>=1.0.0
# Docs: https://pandas.pydata.org/docs/
pandas>=2.2.0,<3

# Excel file support for pandas
# Docs: https://openpyxl.readthedocs.io/
openpyxl>=3.1.0,<4

# Legacy Excel format support (.xls files)
xlrd>=2.0.0,<3

# Microsoft Graph API authentication (Outlook ETL)
msal>=1.24.0

# Office365 REST API Client for SharePoint integration
# Docs: https://github.com/vgrem/office365-rest-python-client
Office365-REST-Python-Client>=2.5.0

# Playwright for HTML to PDF conversion and secure document handling
# Docs: https://playwright.dev/python/
playwright>=1.40.0

gunicorn
whitenoise>=6.6,<7

requests==2.31.0

# Password-protected Excel files
msoffcrypto-tool>=5.0.0

# PDF processing and manipulation
# Docs: https://pypdf.readthedocs.io/
pypdf>=5.1.0,<6
